fix a bug in tagcon when Arg.resolve=False
The filter parser was breaking when encountering unquoted strings containing
dashes when Arg.resolve=False, and seemed to be having some issues with quoted
strings as well; we avoid this by simply not attempting to run filter parsing
when Arg.resolve=False.  This shouldn't break any existing tags, as setting
resolve=False on an Arg means you just want the literal value anyway.